// frontend/generated/models/StringBigDecimalDatapointDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Freelance Platform API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface StringBigDecimalDatapointDto
 */
export interface StringBigDecimalDatapointDto {
    /**
     * 
     * @type {string}
     * @memberof StringBigDecimalDatapointDto
     */
    key?: string;
    /**
     * 
     * @type {number}
     * @memberof StringBigDecimalDatapointDto
     */
    value?: number;
}

/**
 * Check if a given object implements the StringBigDecimalDatapointDto interface.
 */
export function instanceOfStringBigDecimalDatapointDto(value: object): value is StringBigDecimalDatapointDto {
    return true;
}

export function StringBigDecimalDatapointDtoFromJSON(json: any): StringBigDecimalDatapointDto {
    return StringBigDecimalDatapointDtoFromJSONTyped(json, false);
}

export function StringBigDecimalDatapointDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): StringBigDecimalDatapointDto {
    if (json == null) {
        return json;
    }
    return {
        
        'key': json['key'] == null ? undefined : json['key'],
        'value': json['value'] == null ? undefined : json['value'],
    };
}

export function StringBigDecimalDatapointDtoToJSON(json: any): StringBigDecimalDatapointDto {
    return StringBigDecimalDatapointDtoToJSONTyped(json, false);
}

export function StringBigDecimalDatapointDtoToJSONTyped(value?: StringBigDecimalDatapointDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'key': value['key'],
        'value': value['value'],
    };
}