// frontend/generated/models/StringIntegerDatapointDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Freelance Platform API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface StringIntegerDatapointDto
 */
export interface StringIntegerDatapointDto {
    /**
     * 
     * @type {string}
     * @memberof StringIntegerDatapointDto
     */
    key?: string;
    /**
     * 
     * @type {number}
     * @memberof StringIntegerDatapointDto
     */
    value?: number;
}

/**
 * Check if a given object implements the StringIntegerDatapointDto interface.
 */
export function instanceOfStringIntegerDatapointDto(value: object): value is StringIntegerDatapointDto {
    return true;
}

export function StringIntegerDatapointDtoFromJSON(json: any): StringIntegerDatapointDto {
    return StringIntegerDatapointDtoFromJSONTyped(json, false);
}

export function StringIntegerDatapointDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): StringIntegerDatapointDto {
    if (json == null) {
        return json;
    }
    return {
        
        'key': json['key'] == null ? undefined : json['key'],
        'value': json['value'] == null ? undefined : json['value'],
    };
}

export function StringIntegerDatapointDtoToJSON(json: any): StringIntegerDatapointDto {
    return StringIntegerDatapointDtoToJSONTyped(json, false);
}

export function StringIntegerDatapointDtoToJSONTyped(value?: StringIntegerDatapointDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'key': value['key'],
        'value': value['value'],
    };
}