// frontend/generated/models/TransactionStatDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Freelance Platform API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface TransactionStatDto
 */
export interface TransactionStatDto {
    /**
     * Time period (day/week/month) of the data
     * @type {string}
     * @memberof TransactionStatDto
     */
    timePeriod?: string;
    /**
     * Number of successful transactions
     * @type {number}
     * @memberof TransactionStatDto
     */
    successful?: number;
    /**
     * Number of pending transactions
     * @type {number}
     * @memberof TransactionStatDto
     */
    pending?: number;
    /**
     * Number of failed transactions
     * @type {number}
     * @memberof TransactionStatDto
     */
    failed?: number;
}

/**
 * Check if a given object implements the TransactionStatDto interface.
 */
export function instanceOfTransactionStatDto(value: object): value is TransactionStatDto {
    return true;
}

export function TransactionStatDtoFromJSON(json: any): TransactionStatDto {
    return TransactionStatDtoFromJSONTyped(json, false);
}

export function TransactionStatDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): TransactionStatDto {
    if (json == null) {
        return json;
    }
    return {
        
        'timePeriod': json['timePeriod'] == null ? undefined : json['timePeriod'],
        'successful': json['successful'] == null ? undefined : json['successful'],
        'pending': json['pending'] == null ? undefined : json['pending'],
        'failed': json['failed'] == null ? undefined : json['failed'],
    };
}

export function TransactionStatDtoToJSON(json: any): TransactionStatDto {
    return TransactionStatDtoToJSONTyped(json, false);
}

export function TransactionStatDtoToJSONTyped(value?: TransactionStatDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'timePeriod': value['timePeriod'],
        'successful': value['successful'],
        'pending': value['pending'],
        'failed': value['failed'],
    };
}